size context on DropdownMenu
