+++ packages/react/src/components/DropdownMenu/DropdownMenuContext.tsx
@@ -1,8 +1,6 @@
-import { createContext, useState } from 'react';
 import type { ReactNode } from 'react';
 
 import { PopoverContext } from '../Popover';
-import type { DropdownMenuProps } from './DropdownMenu';
 
 export type DropdownMenuContextProps = {
   children: ReactNode;
@@ -22,28 +20,7 @@ export type DropdownMenuContextProps = {
  * </DropdownMenu.Context>
  */
 export const DropdownMenuContext = ({ children }: DropdownMenuContextProps) => {
-  const [size, setSize] =
-    useState<NonNullable<DropdownMenuProps['size']>>('md');
-
-  return (
-    <DropdownMenuCtx.Provider
-      value={{
-        size,
-        setSize,
-      }}
-    >
-      <PopoverContext>{children}</PopoverContext>
-    </DropdownMenuCtx.Provider>
-  );
+  return <PopoverContext>{children}</PopoverContext>;
 };
 
 DropdownMenuContext.displayName = 'DropdownMenuContext';
-
-type DropdownMenuCtxType = {
-  size: NonNullable<DropdownMenuProps['size']>;
-  setSize?: (size: NonNullable<DropdownMenuProps['size']>) => void;
-};
-
-export const DropdownMenuCtx = createContext<DropdownMenuCtxType>({
-  size: 'md',
-});
